image-edit-service/requirements.txt
flask
flask-cors
rembg
Pillow
opencv-python
